prov/efa: split efa_fork_support.c from efa_domain.c and rxr_init.c
This patch move all the functions that implemented EFA's fork support
into efa_fork_support.c.

It also introduced efa_fork_support.h that has declaration of the
functions to use fork support.

It renamed the global variable efa_fork_status to g_efa_fork_status
to conform with the naming convention.
