+++ prov/efa/Makefile.include
@@ -42,6 +42,7 @@ _efa_files = \
 	prov/efa/src/efa_msg.c \
 	prov/efa/src/efa_mr.c \
 	prov/efa/src/efa_rma.c \
+	prov/efa/src/efa_fork_support.c \
 	prov/efa/src/rxr/rxr_attr.c	\
 	prov/efa/src/rxr/rxr_init.c	\
 	prov/efa/src/rxr/rxr_domain.c	\
@@ -61,8 +62,9 @@ _efa_files = \
 
 _efa_headers = \
 	prov/efa/src/efa.h \
-	prov/efa/src/efa_device.h \
 	prov/efa/src/efa_hmem.h \
+	prov/efa/src/efa_device.h \
+	prov/efa/src/efa_fork_support.h \
 	prov/efa/src/rxr/rxr.h \
 	prov/efa/src/rxr/rxr_cntr.h \
 	prov/efa/src/rxr/rxr_rma.h \
